Remove external FontAwesome dependency (#30)
This fixes #28.

Since we only use a single FontAwesome, I decided to just use the SVG (which FontAwesome considers the code comment within appropriate attribution). I further decided "screw it, let's inline it", because it avoids unnecessary complexity, and we can always punt it out to another file later if we need to.

Co-authored-by: Jacob Weisz <inbox@jacobweisz.com>
Reviewed-on: https://git.mills.io/prologic/tube/pulls/30
Co-authored-by: Jacob Weisz <ocdtrekkie@noreply@mills.io>
Co-committed-by: Jacob Weisz <ocdtrekkie@noreply@mills.io>