feat: Add YAML support for added / modified metadata (#37)
This change introduces the ability to read additional
or modified metadata like title or description from a
YAML file.
Just like thumbnails are read from "filename.jpg", we now
look for "filename.yml" and override the meta data that
is embedded in the video file.

This is my very first look at golang, so please tell me if, or rather "where" I went wrong. ;-)

Co-authored-by: Heinrich Langos <gumbo2000@noreply@mills.io>
Reviewed-on: https://git.mills.io/prologic/tube/pulls/37
Co-authored-by: Heinrich 'Henrik' Langos <gumbo2000@noreply@mills.io>
Co-committed-by: Heinrich 'Henrik' Langos <gumbo2000@noreply@mills.io>